Reduced duplication in the FindAndModify operations
JAVA-4544
